GEMINI_ALPHA/requirements.txt
pyTelegramBotAPI
python-dotenv
requests
Pillow
opencv-python
numpy
scikit-image
